+++ src/commands/sqs/mod.rs
@@ -1,1 +1,3 @@
 pub mod models;
+pub mod list_queue;
+pub mod list_message;
